+++ NEWS.md
@@ -1,3 +1,11 @@
+# weektwo 0.0.1
+
+* move sample data in the `test` folder
+* introduce the use of `glue::glue()`
+* introduce some `assertive` checks
+* write tests for all the functions
+* all function in its own files
+
 # weektwo 0.0.0.9000
 
 * Add `inst/sample-data` and move sample data into it